Added objective functions
In SAMCRun() I added a function to calculate the mean of the objective
function over the sampled models. Unfortunately it seems to have large
variance from run to run due to the small number of heavily weighted
samples. (Use the new plotSamples function in utils to see what I mean).
I tried increasing the concentration of the reference density, but it
didn't affect the numbers any. I can't help but wonder if the weighting
factors are incorrect, as I am completely ignoring the log probability
(energy) of the model in calculating the average.

Need to ask Dr. Liang about that.