Working meta-spawner with rsyslog logging
Changed operations so only one py/driver.py would be launched on each
remote machine, and this would then act as a monitoring daemon that
would give heartbeats back to the db, and spawn up children processes as
it deemed necessary. The design is pretty clean, so I'm pretty happy
about it.

I also added the first stages of rsyslog logging in driver.py. Now to
propagate this down into the bowels of SAMC without exploding the
logfiles.

There's still a mystery bug, for some reason the driver will sometimes
take the desired samples below 0 (-1). I'm not sure why this is, but
I've tried a workaround by expanding the ternary on line 104 in
driver.py. Let's see if it works.