Added noisy logic data generation
But now the data is too deterministic (effective sample size is
something like 1/100), so I need to increase the noise. This is starting
to show the advantage of templates.

I also added a plotScatter function which is useful when looking at
WHICH samples will be used most heavily at which energies.