Working status on the night before meeting with Ed
Doing a lot of overnight simulations, but things actually appear to be
working now. Not quite as well as you'd hope, but perhaps I was too
ambitious. Now to see what the data looks like in the morning and
massage it into more beautiful forms.

Lots of small changes, most notably adding more configuration options
for noisy data, experiment types, data method generation (or was that
last commit?).

Also removed pytables HD5 disk sample saving. It was just making
everything overly complicated in that there was always a check to see
what kind of database was there, and then maintaining those two code
paths across the two classes.

I also moved the prior structure calculation to the bottom of cost.c. I
think it is good there as it makes it nice and simple (and working).

I also rehauled the entire database into a simple 3-column numpy array
for maximal memory efficiency. It also precomputes the function which is
now hardcoded into the bayesnet class. It slows down the SAMC sampling
by around 50% but we would have done the calculation at the end anyway
(most of the time). I still wonder if there's a faster way to do this,
probably with a fast C loop that uses the x[] translation table to
compare the two matrices and add up the sum.