Completely working overhaul with libdai
Now using libdai for marginal calculation, efficient KLD calculation, and
energy calculations. Performance appears to be okay, with a slightly greater
than linear slowdown on increasing the number of nodes.

No profiling has been done yet, so I don't know why it is slow. In rewriting, I
fixed a *showstopper* bug in the energy calculation that was giving wrong
answers (I still am in disbelief, it was in setting self.fvalue[i] = accum,
which i was the wrong index to be using, it should have been m or node).

Acceptance is much better now that that is fixed so now to run some large scale
tests to test this.