Updated tail to use proper priors
Found out that there is a few missing constants in the analytic solution
of tail.py effective densities (due to differences between inverse
wishart and gamma). I still need to find the values of the actual
constants, but as can be seen by the ludicrously close match between the
gavg and sampled predictive posterior, the method is working correctly.