Too many things to mention...
Increased the number of variables input into init of mpmDist, including
parameters mumove and lammove to adjust acceptance rate.

Turned back on lambda sampling, and allow d to be passed into mpmDist.

Have a good test case for mixture modeling in cov_poisson, but currently the
resulting g_avg looks wrong. I think this is because the order of summations is
wrong in calc_gavg, but I'm going to fix it tomorrow.

Added scheme_propose and scheme_accept to get a better feel for the types of
proposals and acceptance moves we are making in a run.